requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-heroku==0.3.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
phonenumbers==8.12.18
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
sentry-sdk==0.20.3
twilio==6.52.0
